refactor- folder for related components and naming
