+++ weather-app-frontend/src/layout/Header.tsx
@@ -1,12 +1,12 @@
 import { ReactNode } from "react";
 import { Flex, Heading } from "@chakra-ui/react";
-import Navbar from "./Navbar";
+import Navbar from "../components/navbar/Navbar";
 
-interface HeaderProps {
+interface RootLayoutHeaderProps {
   children: ReactNode;
 }
 
-const Header: React.FC<HeaderProps> = ({ children }) => {
+const RootLayoutHeader: React.FC<RootLayoutHeaderProps> = ({ children }) => {
   return (
     <>
       <Flex
@@ -23,4 +23,4 @@ const Header: React.FC<HeaderProps> = ({ children }) => {
   );
 };
 
-export default Header;
+export default RootLayoutHeader;
